pyproject.toml
[project]
name = "readcast"
version = "0.1.0"
description = "Turn articles into podcasts on your Mac"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Greg" }
]
keywords = ["tts", "podcast", "articles", "offline", "apple-silicon", "kokoro"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Environment :: Web Environment",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS :: MacOS X",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
  "httpx>=0.27",
  "readability-lxml>=0.8",
  "beautifulsoup4>=4.12",
  "lxml>=5.0",
  "mutagen>=1.47",
  "click>=8.1",
  "rich>=13.0",
  "fastapi>=0.115",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "build>=1.2",
  "hatchling>=1.27",
]

[project.scripts]
readcast = "readcast.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
  "src/readcast/**/*.py",
  "src/readcast/web/static/*.html",
  "src/readcast/web/static/*.js",
]

[tool.hatch.build.targets.wheel]
packages = ["src/readcast"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
markers = [
  "integration: requires local runtime dependencies such as ffmpeg and kokoro-edge",
]

[tool.ruff]
target-version = "py311"
line-length = 140
exclude = [".pixi", ".pytest_cache", ".venv", "node_modules", "dist", "build"]

[tool.ruff.lint]
select = ["E", "F"]
